+++ frontend/appflowy_flutter/lib/plugins/database_view/board/application/card/board_number_cell_bloc.dart
@@ -1,7 +1,8 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'dart:async';
-import '../../../application/cell/cell_service.dart';
+
+import '../../../application/cell/cell_controller_builder.dart';
 
 part 'board_number_cell_bloc.freezed.dart';
