+++ frontend/appflowy_flutter/lib/plugins/database_view/board/application/card/card_data_controller.dart
@@ -1,9 +1,7 @@
 import 'package:appflowy_backend/protobuf/flowy-database/row_entities.pb.dart';
 import 'package:flutter/foundation.dart';
 
-import '../../../application/cell/cell_field_notifier.dart';
 import '../../../application/cell/cell_service.dart';
-import '../../../application/field/field_controller.dart';
 import '../../../application/row/row_cache.dart';
 import '../../presentation/card/card_cell_builder.dart';
 
@@ -11,16 +9,13 @@ typedef OnCardChanged = void Function(CellByFieldId, RowsChangedReason);
 
 class CardDataController extends BoardCellBuilderDelegate {
   final RowPB rowPB;
-  final FieldController _fieldController;
   final RowCache _rowCache;
   final List<VoidCallback> _onCardChangedListeners = [];
 
   CardDataController({
     required this.rowPB,
-    required FieldController fieldController,
     required RowCache rowCache,
-  })  : _fieldController = fieldController,
-        _rowCache = rowCache;
+  }) : _rowCache = rowCache;
 
   CellByFieldId loadData() {
     return _rowCache.loadGridCells(rowPB.id);
@@ -37,12 +32,6 @@ class CardDataController extends BoardCellBuilderDelegate {
     for (final fn in _onCardChangedListeners) {
       _rowCache.removeRowListener(fn);
     }
-  }
-
-  @override
-  CellFieldNotifier buildFieldNotifier() {
-    return CellFieldNotifier(
-        notifier: GridCellFieldNotifierImpl(_fieldController));
   }
 
   @override
