+++ frontend/appflowy_flutter/lib/plugins/database_view/board/presentation/card/board_checklist_cell.dart
@@ -1,7 +1,7 @@
+import 'package:appflowy/plugins/database_view/application/cell/cell_controller_builder.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../application/cell/cell_service.dart';
 import '../../../grid/application/cell/checklist_cell_bloc.dart';
 import '../../../grid/presentation/widgets/cell/checklist_cell/checklist_progress_bar.dart';
