+++ frontend/appflowy_flutter/lib/plugins/database_view/board/presentation/card/board_date_cell.dart
@@ -1,9 +1,9 @@
+import 'package:appflowy/plugins/database_view/application/cell/cell_controller_builder.dart';
 import 'package:appflowy/plugins/database_view/board/application/card/board_date_cell_bloc.dart';
 import 'package:flowy_infra_ui/style_widget/text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../application/cell/cell_service.dart';
 import 'define.dart';
 
 class BoardDateCell extends StatefulWidget {
