+++ frontend/appflowy_flutter/lib/plugins/database_view/board/presentation/card/board_select_option_cell.dart
@@ -1,9 +1,8 @@
+import 'package:appflowy/plugins/database_view/application/cell/cell_controller_builder.dart';
 import 'package:appflowy_popover/appflowy_popover.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
-import '../../../application/cell/cell_service.dart';
 import '../../../grid/presentation/widgets/cell/select_option_cell/extension.dart';
 import '../../../grid/presentation/widgets/cell/select_option_cell/select_option_editor.dart';
 import '../../application/card/board_select_option_cell_bloc.dart';
