+++ frontend/appflowy_flutter/lib/plugins/database_view/board/presentation/card/board_url_cell.dart
@@ -1,9 +1,9 @@
+import 'package:appflowy/plugins/database_view/application/cell/cell_controller_builder.dart';
 import 'package:flowy_infra/size.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:textstyle_extensions/textstyle_extensions.dart';
 
-import '../../../application/cell/cell_service.dart';
 import '../../application/card/board_url_cell_bloc.dart';
 import 'define.dart';
