+++ frontend/appflowy_flutter/lib/startup/deps_resolver.dart
@@ -1,4 +1,5 @@
 import 'package:appflowy/core/network_monitor.dart';
+import 'package:appflowy/plugins/database_view/application/cell/cell_controller_builder.dart';
 import 'package:appflowy/plugins/database_view/application/cell/cell_service.dart';
 import 'package:appflowy/plugins/database_view/application/field/field_action_sheet_bloc.dart';
 import 'package:appflowy/plugins/database_view/application/field/field_controller.dart';
